refactor: move `document_to_element_list` from `file_utils/filetype.py` to `partition/common.py` to fix an issue with importing `numpy`
