+++ unstructured/partition/pdf.py
@@ -24,7 +24,6 @@ from unstructured.documents.elements import (
 from unstructured.file_utils.filetype import (
     FileType,
     add_metadata_with_filetype,
-    document_to_element_list,
 )
 from unstructured.nlp.patterns import PARAGRAPH_PATTERN
 from unstructured.partition.common import (
@@ -33,6 +32,7 @@ from unstructured.partition.common import (
     get_last_modified_date,
     get_last_modified_date_from_file,
     spooled_to_bytes_io_if_needed,
+    document_to_element_list,
 )
 from unstructured.partition.strategies import determine_pdf_or_image_strategy
 from unstructured.partition.text import element_from_text, partition_text
